Time: 15 ms (98.35%), Space: 16.3 MB (43.04%) - LeetHub

+++ 0155-min-stack/0155-min-stack.cpp
@@ -53,7 +53,7 @@ public:
         return minEle;
     }
     
-    int getMin() {
+    long long getMin() {
         if(s.size()==0)
             return -1;
         return minEle;
